fixed rotate operators

+++ Test/Expect/test17.cpp
@@ -7,6 +7,6 @@ uint8_t test17::dat[] = {
 int32_t test17::start(void)
 {
   int32_t result = 0;
-  return (&(*(int32_t *)&dat[0]));
+  return ((int32_t)&(*(int32_t *)&dat[0]));
 }
